requirements.txt
# https://toga.readthedocs.io/en/stable/tutorial/tutorial-0.html
toga
mysql-connector-python